Try removing the parameter from the iFrame directly
